test(film): add test for findAll method
Difficulties encountered while trying to cass the `findAll` method.
`this.store().findAll('film')` would just call `/films` instead of the url defined in the app's custom adapter.
Had to add `needs: ['adapter:application', 'serializer:application']` inside the test's configuration to make it aware of both the url and the serializer.